+++ example/lib/views.dart
@@ -1,3 +1,0 @@
-export 'views/home_view.dart';
-export 'views/gatt_view.dart';
-export 'views/flip_view.dart';