+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Fix connect blocked when bluetooth closed.
+- Fix wrong repository url.
+- Move all example files to main.dart.
+
 ## 0.0.1
 
 - Add central APIs.